+++ lib/sync/index.js
@@ -7,7 +7,7 @@ var metricsModule = require('./sync-metrics');
 var MongodbQueue = require('./mongodbQueue');
 var Worker = require('./worker');
 var dataHandlersModule = require('./dataHandlers');
-var cacheModule = require('./cache');
+var storageModule = require('./storage');
 var hashProviderModule = require('./hashProvider');
 var ackProcessor = require('./ack-processor');
 var pendingProcessor = require('./pending-processor');
@@ -20,7 +20,7 @@ var redisClient = null;
 var mongoDbClient = null;
 var metricsClient = null;
 var dataHandlers = null;
-var syncCache = null;
+var syncStorage = null;
 var hashProvider = null;
 
 var syncStarted = false;
@@ -186,7 +186,7 @@ function start(cb) {
   }
 
   dataHandlers = dataHandlersModule();
-  syncCache = cacheModule(redisClient);
+  syncStorage = storageModule(redisClient);
   hashProvider = hashProviderModule();
 
   async.waterfall([
@@ -202,7 +202,7 @@ function start(cb) {
       ], callback);
     },
     function createWorkers(callback) {
-      var syncProcessorImpl = syncProcessor(syncCache, dataHandlers, metricsClient, hashProvider);
+      var syncProcessorImpl = syncProcessor(syncStorage, dataHandlers, metricsClient, hashProvider);
       syncWorker = new Worker(syncQueue, syncProcessorImpl, metricsClient, {name: 'sync_worker'});
 
       ackWorker = new Worker(ackQueue, ackProcessor, metricsClient, {name: 'ack_worker'});
