+++ lib/sync/sync-processor.js
@@ -2,7 +2,7 @@ var util = require('util');
 var _ = require('underscore');
 var syncUtil = require('./util');
 
-var syncCache, dataHandlers, metricsClient, hashProvider;
+var syncStorage, dataHandlers, metricsClient, hashProvider;
 
 /**
  * Perform the sync operation
@@ -48,7 +48,7 @@ function syncWithBackend(payload, callback) {
       });
       var globalHash = hashProvider.globalHash(datasetId, _.pluck(recordsWithHash, 'hash'));
       syncInfo.syncLoopEnd = new Date().getTime();
-      syncCache.saveLastSyncDataset(datasetId, globalHash, recordsWithHash, syncInfo, function(err){
+      syncStorage.saveLastSyncDataset(datasetId, globalHash, recordsWithHash, syncInfo, function(err){
         if (err) {
           syncUtil.doLog(datasetId, "error", "error when save data to cache. error = " + util.inspect(err));
         }
@@ -59,8 +59,8 @@ function syncWithBackend(payload, callback) {
   });
 }
 
-module.exports = function(syncCacheImpl, dataHandlersImpl, metricsClientImpl, hashProviderImpl) {
-  syncCache = syncCacheImpl;
+module.exports = function(syncStorageImpl, dataHandlersImpl, metricsClientImpl, hashProviderImpl) {
+  syncStorage = syncStorageImpl;
   dataHandlers = dataHandlersImpl;
   metricsClient = metricsClientImpl;
   hashProvider = hashProviderImpl;
